use AN coding only with 8-bit A for now since we do not have code handling data bit-/byte-wise. This allows to use only 16 bits for resilient tinyint, 32 bits for resilient short (only 24 needed) and 64 bits for resilient int (onyl 40 neeeded). Fix output in lineorder_size uncomment 16-bit A's

+++ include/ColumnStore.h
@@ -64,10 +64,11 @@ typedef uint8_t tinyint_t;
 typedef uint16_t shortint_t;
 typedef uint32_t int_t;
 typedef uint64_t largeint_t;
+
 typedef char char_t, *str_t;
 typedef double fixed_t;
 
-typedef uint32_t restiny_t;
+typedef uint16_t restiny_t;
 typedef uint32_t resshort_t;
 typedef uint64_t resint_t;
